Engine: Prepare for LRU with linked list re-using same value of map
Now while we store every value on a map, we also re-use same entry for
building a linked list where last used (get/set) is moved at the end
(tail). In this way it is more easy to implement LRU based on how many
items we want store, or based on size (Mb).

Signed-off-by: Federico Guerinoni <guerinoni.federico@gmail.com>